final_concur_reports_processor/requirements.txt
openai>=1.30.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
